+++ CHANGELOG.md
@@ -44,6 +44,8 @@ high state of flux, you're at risk of it changing without notice.
     - make `intercalate` pipeable (@gcanti)
   - `IOEither`
     - remove `getIOValidation`, use `getApplicativeIOValidation` and `getAltIOValidation` instead (@gcanti)
+  - `Monoid`
+    - swap execution order in `getEndomorphismMonoid` (@gcanti)
 
 # 2.9.1